pyproject.toml
[tool.poetry]
name = "blocking"
version = "0.1.0"
description = ""
authors = ["tobi <mail.tobias.leibrock@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9,<3.13"
numpy = "^1.26.1"
netCDF4 = "^1.6.5"
matplotlib = "^3.8.1"
Cartopy = "^0.22.0"
ipykernel = "^6.26.0"
scipy = "^1.11.3"
ipympl = "^0.9.3"
ipywidgets = "^8.1.1"
nc-time-axis = "^1.4.1"
seaborn = "^0.13.0"
xarray = "^2023.12.0"
torch = {version = "^2.1.2+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.16.2+cu118", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.1.2+cu118", source = "pytorch-gpu-src"}
torchmetrics = "^1.2.1"
scikit-learn = "^1.3.2"
tensorboard = "^2.15.1"
pandas = "^2.1.4"
pillow = "^10.2.0"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
